+++ Phoenix/PHModalWindowController.h
@@ -31,6 +31,10 @@
 
 - (instancetype) init;
 
+#pragma mark - Appearance
+
+JSExportAs(setTextColor, - (void) setTextColorWithRed:(CGFloat)red green:(CGFloat)green blue:(CGFloat)blue alpha:(CGFloat)alpha);
+
 #pragma mark - Displaying
 
 - (NSRect) frame;